# dsd: pool metrics.MetricSample