--- a/cmd/dogstatsd/main.go
+++ b/cmd/dogstatsd/main.go
@@ -17,6 +17,8 @@ import (
 	"os/signal"
 	"syscall"
 
+	"github.com/DataDog/datadog-agent/pkg/metrics"
+
 	"github.com/spf13/cobra"
 
 	"github.com/DataDog/datadog-agent/pkg/aggregator"
@@ -182,8 +184,10 @@ func start(cmd *cobra.Command, args []string) error {
 		tagger.Init()
 	}
 
+	metricSamplePool := metrics.NewMetricSamplePool(32)
 	aggregatorInstance := aggregator.InitAggregator(s, hname, "agent")
-	statsd, err := dogstatsd.NewServer(aggregatorInstance.GetBufferedChannels())
+	sampleC, eventC, serviceCheckC := aggregatorInstance.GetBufferedChannels()
+	statsd, err := dogstatsd.NewServer(metricSamplePool, sampleC, eventC, serviceCheckC)
 	if err != nil {
 		log.Criticalf("Unable to start dogstatsd: %s", err)
 		return nil