--- a/pkg/dogstatsd/server.go
+++ b/pkg/dogstatsd/server.go
@@ -18,13 +18,14 @@ import (
 	"sync"
 	"time"
 
+	"github.com/DataDog/datadog-agent/pkg/tagger"
+
 	"github.com/DataDog/datadog-agent/pkg/util/log"
 
 	"github.com/DataDog/datadog-agent/pkg/config"
 	"github.com/DataDog/datadog-agent/pkg/dogstatsd/listeners"
 	"github.com/DataDog/datadog-agent/pkg/metrics"
 	"github.com/DataDog/datadog-agent/pkg/status/health"
-	"github.com/DataDog/datadog-agent/pkg/tagger"
 	"github.com/DataDog/datadog-agent/pkg/util"
 )
 
@@ -53,6 +54,9 @@ type Server struct {
 	listeners             []listeners.StatsdListener
 	packetsIn             chan listeners.Packets
 	samplePool            *metrics.MetricSamplePool
+	samplesOut            chan<- []*metrics.MetricSample
+	eventsOut             chan<- []*metrics.Event
+	servicesCheckOut      chan<- []*metrics.ServiceCheck
 	Statistics            *util.Stats
 	Started               bool
 	packetPool            *listeners.PacketPool
@@ -77,7 +81,7 @@ type metricStat struct {
 }
 
 // NewServer returns a running Dogstatsd server
-func NewServer(samplePool *metrics.MetricSamplePool, metricOut chan<- []*metrics.MetricSample, eventOut chan<- []*metrics.Event, serviceCheckOut chan<- []*metrics.ServiceCheck) (*Server, error) {
+func NewServer(samplePool *metrics.MetricSamplePool, samplesOut chan<- []*metrics.MetricSample, eventsOut chan<- []*metrics.Event, servicesCheckOut chan<- []*metrics.ServiceCheck) (*Server, error) {
 	var stats *util.Stats
 	if config.Datadog.GetBool("dogstatsd_stats_enable") == true {
 		buff := config.Datadog.GetInt("dogstatsd_stats_buffer")
@@ -143,6 +147,9 @@ func NewServer(samplePool *metrics.MetricSamplePool, metricOut chan<- []*metrics
 		Statistics:            stats,
 		samplePool:            samplePool,
 		packetsIn:             packetsChannel,
+		samplesOut:            samplesOut,
+		eventsOut:             eventsOut,
+		servicesCheckOut:      servicesCheckOut,
 		listeners:             tmpListeners,
 		packetPool:            packetPool,
 		stopChan:              make(chan bool),
@@ -173,13 +180,11 @@ func NewServer(samplePool *metrics.MetricSamplePool, metricOut chan<- []*metrics
 			go s.forwarder(con, packetsChannel)
 		}
 	}
-
-	s.handleMessages(metricOut, eventOut, serviceCheckOut)
-
+	s.handleMessages()
 	return s, nil
 }
 
-func (s *Server) handleMessages(metricOut chan<- []*metrics.MetricSample, eventOut chan<- []*metrics.Event, serviceCheckOut chan<- []*metrics.ServiceCheck) {
+func (s *Server) handleMessages() {
 	if s.Statistics != nil {
 		go s.Statistics.Process()
 		go s.Statistics.Update(&dogstatsdPacketsLastSec)
@@ -197,7 +202,7 @@ func (s *Server) handleMessages(metricOut chan<- []*metrics.MetricSample, eventO
 	}
 
 	for i := 0; i < workers; i++ {
-		go s.worker(metricOut, eventOut, serviceCheckOut)
+		go s.worker()
 	}
 }
 
@@ -219,31 +224,14 @@ func (s *Server) forwarder(fcon net.Conn, packetsChannel chan listeners.Packets)
 	}
 }
 
-func (s *Server) worker(metricOut chan<- []*metrics.MetricSample, eventOut chan<- []*metrics.Event, serviceCheckOut chan<- []*metrics.ServiceCheck) {
+func (s *Server) worker() {
 	for {
 		select {
 		case <-s.stopChan:
 			return
 		case <-s.health.C:
 		case packets := <-s.packetsIn:
-			events := make([]*metrics.Event, 0, len(packets))
-			serviceChecks := make([]*metrics.ServiceCheck, 0, len(packets))
-			metricSamples := make([]*metrics.MetricSample, 0, len(packets))
-
-			for _, packet := range packets {
-				metricSamples, events, serviceChecks = s.parsePacket(packet, metricSamples, events, serviceChecks)
-				s.packetPool.Put(packet)
-			}
-
-			if len(metricSamples) != 0 {
-				metricOut <- metricSamples
-			}
-			if len(events) != 0 {
-				eventOut <- events
-			}
-			if len(serviceChecks) != 0 {
-				serviceCheckOut <- serviceChecks
-			}
+			s.parsePackets(packets)
 		}
 	}
 }
@@ -262,77 +250,119 @@ func nextMessage(packet *[]byte) (message []byte) {
 	return message
 }
 
-func (s *Server) parsePacket(packet *listeners.Packet, metricSamples []*metrics.MetricSample, events []*metrics.Event, serviceChecks []*metrics.ServiceCheck) ([]*metrics.MetricSample, []*metrics.Event, []*metrics.ServiceCheck) {
-	extraTags := s.extraTags
+func (s *Server) parsePackets(packets []*listeners.Packet) {
+	var events []*metrics.Event
+	var serviceChecks []*metrics.ServiceCheck
+	var metricSamples []*metrics.MetricSample
 
-	log.Tracef("Dogstatsd receive: %s", packet.Contents)
-	if packet.Origin != listeners.NoOrigin {
-		originTags, err := tagger.Tag(packet.Origin, tagger.DogstatsdCardinality)
+	for _, packet := range packets {
+		originTags := findOriginTags(packet.Origin)
+		log.Tracef("Dogstatsd receive: %s", packet.Contents)
+		for {
+			message := nextMessage(&packet.Contents)
+			if message == nil {
+				break
+			}
+			if s.Statistics != nil {
+				s.Statistics.StatEvent(1)
+			}
+			messageType := findMessageType(message)
+
+			switch messageType {
+			case serviceCheckType:
+				serviceCheck, err := s.parseServiceCheckMessage(message)
+				if err != nil {
+					log.Errorf("Dogstatsd: error parsing service check: %s", err)
+					continue
+				}
+				serviceCheck.Tags = append(serviceCheck.Tags, originTags...)
+				serviceChecks = append(serviceChecks, serviceCheck)
+			case eventType:
+				event, err := s.parseEventMessage(message)
+				if err != nil {
+					log.Errorf("Dogstatsd: error parsing event: %s", err)
+					continue
+				}
+				event.Tags = append(event.Tags, originTags...)
+				events = append(events, event)
+			case metricSampleType:
+				sample, err := s.parseMetricMessage(message)
+				if err != nil {
+					log.Errorf("Dogstatsd: error parsing metrics: %s", err)
+					continue
+				}
+				if s.debugMetricsStats {
+					s.storeMetricStats(sample.Name)
+				}
+				sample.Tags = append(sample.Tags, originTags...)
+				metricSamples = append(metricSamples, sample)
+				if s.histToDist && sample.Mtype == metrics.HistogramType {
+					distSample := sample.Copy()
+					distSample.Name = s.histToDistPrefix + distSample.Name
+					distSample.Mtype = metrics.DistributionType
+					metricSamples = append(metricSamples, distSample)
+				}
+			}
+		}
+	}
+	if len(serviceChecks) > 0 {
+		s.servicesCheckOut <- serviceChecks
+	}
+	if len(events) > 0 {
+		s.eventsOut <- events
+	}
+	if len(metricSamples) > 0 {
+		s.samplesOut <- metricSamples
+	}
+}
+
+func (s *Server) parseMetricMessage(message []byte) (*metrics.MetricSample, error) {
+	sample, err := parseMetricSample(message)
+	if err != nil {
+		dogstatsdMetricParseErrors.Add(1)
+		return nil, err
+	}
+	metricSample := enrichMetricSample(sample, s.metricPrefix, s.metricPrefixBlacklist, s.defaultHostname)
+	metricSample.Tags = append(metricSample.Tags, s.extraTags...)
+	dogstatsdMetricPackets.Add(1)
+	return metricSample, nil
+}
+
+func (s *Server) parseEventMessage(message []byte) (*metrics.Event, error) {
+	sample, err := parseEvent(message)
+	if err != nil {
+		dogstatsdEventParseErrors.Add(1)
+		return nil, err
+	}
+	event := enrichEvent(sample, s.defaultHostname)
+	event.Tags = append(event.Tags, s.extraTags...)
+	dogstatsdEventPackets.Add(1)
+	return event, nil
+}
+
+func (s *Server) parseServiceCheckMessage(message []byte) (*metrics.ServiceCheck, error) {
+	sample, err := parseServiceCheck(message)
+	if err != nil {
+		dogstatsdServiceCheckParseErrors.Add(1)
+		return nil, err
+	}
+	serviceCheck := enrichServiceCheck(sample, s.defaultHostname)
+	serviceCheck.Tags = append(serviceCheck.Tags, s.extraTags...)
+	dogstatsdServiceCheckPackets.Add(1)
+	return serviceCheck, nil
+}
+
+func findOriginTags(origin string) []string {
+	var tags []string
+	if origin != listeners.NoOrigin {
+		originTags, err := tagger.Tag(origin, tagger.DogstatsdCardinality)
 		if err != nil {
 			log.Errorf(err.Error())
 		} else {
-			extraTags = append(extraTags, originTags...)
+			tags = append(tags, originTags...)
 		}
 	}
-
-	for {
-		message := nextMessage(&packet.Contents)
-		if message == nil {
-			break
-		}
-
-		if s.Statistics != nil {
-			s.Statistics.StatEvent(1)
-		}
-
-		if bytes.HasPrefix(message, []byte("_sc")) {
-			serviceCheck, err := parseServiceCheckMessage(message, s.defaultHostname)
-			if err != nil {
-				log.Errorf("Dogstatsd: error parsing service check: %s", err)
-				dogstatsdServiceCheckParseErrors.Add(1)
-				continue
-			}
-			if len(extraTags) > 0 {
-				serviceCheck.Tags = append(serviceCheck.Tags, extraTags...)
-			}
-			dogstatsdServiceCheckPackets.Add(1)
-			serviceChecks = append(serviceChecks, serviceCheck)
-		} else if bytes.HasPrefix(message, []byte("_e")) {
-			event, err := parseEventMessage(message, s.defaultHostname)
-			if err != nil {
-				log.Errorf("Dogstatsd: error parsing event: %s", err)
-				dogstatsdEventParseErrors.Add(1)
-				continue
-			}
-			if len(extraTags) > 0 {
-				event.Tags = append(event.Tags, extraTags...)
-			}
-			dogstatsdEventPackets.Add(1)
-			events = append(events, event)
-		} else {
-			sample, err := parseMetricMessage(message, s.metricPrefix, s.metricPrefixBlacklist, s.defaultHostname)
-			if err != nil {
-				log.Errorf("Dogstatsd: error parsing metrics: %s", err)
-				dogstatsdMetricParseErrors.Add(1)
-				continue
-			}
-			if s.debugMetricsStats {
-				s.storeMetricStats(sample.Name)
-			}
-			if len(extraTags) > 0 {
-				sample.Tags = append(sample.Tags, extraTags...)
-			}
-			dogstatsdMetricPackets.Add(1)
-			metricSamples = append(metricSamples, sample)
-			if s.histToDist && sample.Mtype == metrics.HistogramType {
-				distSample := sample.Copy()
-				distSample.Name = s.histToDistPrefix + distSample.Name
-				distSample.Mtype = metrics.DistributionType
-				metricSamples = append(metricSamples, distSample)
-			}
-		}
-	}
-	return metricSamples, events, serviceChecks
+	return tags
 }
 
 // Stop stops a running Dogstatsd server