--- a/pkg/metrics/metric_sample_pool.go
+++ b/pkg/metrics/metric_sample_pool.go
@@ -1,0 +1,32 @@
+package metrics
+
+import (
+	"sync"
+)
+
+// MetricSamplePool is a pool of metrics sample
+type MetricSamplePool struct {
+	pool *sync.Pool
+}
+
+// NewMetricSamplePool creates a new MetricSamplePool
+func NewMetricSamplePool(batchSize int) *MetricSamplePool {
+	return &MetricSamplePool{
+		pool: &sync.Pool{
+			New: func() interface{} {
+				batch := make([]MetricSample, 0, batchSize)
+				return &batch
+			},
+		},
+	}
+}
+
+// GetBatch gets a batch of metric samples from the pool
+func (m *MetricSamplePool) GetBatch() []MetricSample {
+	return m.pool.Get().([]MetricSample)
+}
+
+// PutBatch puts a batch back into the pool
+func (m *MetricSamplePool) PutBatch(batch []MetricSample) {
+	m.pool.Put(batch[:0])
+}