### What does this PR do?

Create pool of `metrics.MetricSample` batches shared between dogstatsd parsers and the aggregator. 

### Motivation

It's currently expensive to store one object on the heap for each incoming sample. This PR brings two benefits:

1. Reduce the number of pointers the GC has to keep track of by allocating slices of samples instead of slice of pointers to samples.
2. Re-use the number and size of allocations by using a `sync.Pool`

### Benchmark

Parsing 640 sample:

```
benchmark                   old ns/op     new ns/op     delta
BenchmarkParsePackets-4     246904        192589        -22.00%

benchmark                   old allocs     new allocs     delta
BenchmarkParsePackets-4     4495           3243           -27.85%

benchmark                   old bytes     new bytes     delta
BenchmarkParsePackets-4     170349        70636         -58.53%
```

### Notes

This PR targets https://github.com/DataDog/datadog-agent/pull/4575